refactored pub-sub functions into client; doc added[in README]
